NavigationBarView: make sure we can't get stuck with the keyboard open
The layout changers turn into keyboard controls with multiple layouts and if
we're on a layout without a back key and no hardware keys this can leave us
stuck with the keyboard open.  This will have to be redone at some point to
better handle all the keyboard states.

Change-Id: I4dca7023c5ada30e176d5fda578f28143b7cbce0

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NavigationBarView.java